refactor - move utils into seperate part - update comments
