+++ nerif/agent/__init__.py
@@ -1,10 +1,7 @@
-from .agent import (
-    LogitsAgent,
+from .agent import LogitsAgent, SimpleChatAgent, SimpleEmbeddingAgent, VisionAgent
+from .token_counter import NerifTokenCounter
+from .utils import (
     MessageType,
-    NerifTokenCounter,
-    SimpleChatAgent,
-    SimpleEmbeddingAgent,
-    VisionAgent,
     get_litellm_embedding,
     get_litellm_response,
     get_ollama_response,
